fix(frame-handle): satisfy pre-push clang-tidy checks

+++ playback/output/outputframecache.h
@@ -18,7 +18,7 @@ public:
     void insertAudioFrame(const MediaAudioFrame& frame);
     QByteArray audioSpanOrSilence(int feedIndex, qint64 startSample, int sampleFrames) const;
 
-    int feedCount() const { return m_video.size(); }
+    int feedCount() const { return static_cast<int>(m_video.size()); }
     // Insert every frame from `other` (video + audio) without removing the
     // current contents. Feed counts must match. Used to merge a staging window
     // into the live cache before trimming old frames (double-buffer).
